feat: upgrade to flyway 6
