+++ extensions/flyway/runtime/src/main/java/io/quarkus/flyway/runtime/graal/CompositeMigrationResolverSubstitutions.java
@@ -7,12 +7,12 @@ import java.util.Collection;
 import org.flywaydb.core.api.configuration.Configuration;
 import org.flywaydb.core.api.resolver.MigrationResolver;
 import org.flywaydb.core.internal.clazz.ClassProvider;
-import org.flywaydb.core.internal.database.base.Database;
-import org.flywaydb.core.internal.resolver.java.JavaMigrationResolver;
-import org.flywaydb.core.internal.resolver.jdbc.JdbcMigrationResolver;
+import org.flywaydb.core.internal.resolver.java.FixedJavaMigrationResolver;
+import org.flywaydb.core.internal.resolver.java.ScanningJavaMigrationResolver;
 import org.flywaydb.core.internal.resolver.sql.SqlMigrationResolver;
 import org.flywaydb.core.internal.resource.ResourceProvider;
-import org.flywaydb.core.internal.sqlscript.SqlStatementBuilderFactory;
+import org.flywaydb.core.internal.sqlscript.SqlScriptExecutorFactory;
+import org.flywaydb.core.internal.sqlscript.SqlScriptFactory;
 
 import com.oracle.svm.core.annotate.Alias;
 import com.oracle.svm.core.annotate.Substitute;
@@ -29,26 +29,24 @@ public final class CompositeMigrationResolverSubstitutions {
      * in the resolvers list to avoid native image errors because of the incomplete classpath
      *
      * @see org.flywaydb.core.internal.resolver.spring.SpringJdbcMigrationResolver
-     * @see org.flywaydb.core.internal.resolver.CompositeMigrationResolver#CompositeMigrationResolver(Database,
-     *      ResourceProvider, ClassProvider, Configuration, SqlStatementBuilderFactory, MigrationResolver...)
+     * @see org.flywaydb.core.internal.resolver.CompositeMigrationResolver#CompositeMigrationResolver(ResourceProvider,
+     *      ClassProvider, Configuration, SqlScriptExecutorFactory, SqlScriptFactory, MigrationResolver...)
      */
     @Substitute
-    public CompositeMigrationResolverSubstitutions(
-            Database database,
-            ResourceProvider resourceProvider,
+    public CompositeMigrationResolverSubstitutions(ResourceProvider resourceProvider,
             ClassProvider classProvider,
             Configuration configuration,
-            SqlStatementBuilderFactory sqlStatementBuilderFactory,
+            SqlScriptExecutorFactory sqlScriptExecutorFactory,
+            SqlScriptFactory sqlScriptFactory,
             MigrationResolver... customMigrationResolvers) {
         if (!configuration.isSkipDefaultResolvers()) {
-            migrationResolvers.add(new SqlMigrationResolver(
-                    database,
-                    resourceProvider,
-                    sqlStatementBuilderFactory,
+            migrationResolvers.add(new SqlMigrationResolver(resourceProvider, sqlScriptExecutorFactory, sqlScriptFactory,
                     configuration));
-            migrationResolvers.add(new JavaMigrationResolver(classProvider, configuration));
-            migrationResolvers.add(new JdbcMigrationResolver(classProvider, configuration));
+            migrationResolvers.add(new ScanningJavaMigrationResolver(classProvider, configuration));
         }
+        migrationResolvers.add(new FixedJavaMigrationResolver(configuration.getJavaMigrations()));
+
         migrationResolvers.addAll(Arrays.asList(customMigrationResolvers));
     }
+
 }
